Seperate RunTimer and RunDMA into dedicated codefiles.
