+++ src/gba/memory.cpp
@@ -153,152 +153,6 @@ namespace NanoboyAdvance
         delete m_Video;
         delete m_Interrupt;
     }
-    
-
-    ///////////////////////////////////////////////////////////
-    /// \author Frederic Meyer
-    /// \date   July 31th, 2016
-    /// \fn     RunTimer
-    ///
-    ///////////////////////////////////////////////////////////
-    void GBAMemory::RunTimer()
-    {
-        bool overflow = false;
-        
-        for (int i = 0; i < 4; i++)
-        {
-            if (!m_Timer[i].enable) continue;
-
-            if ((m_Timer[i].countup && overflow) || 
-                (!m_Timer[i].countup && ++m_Timer[i].ticks >= tmr_cycles[m_Timer[i].clock]))
-            {
-                m_Timer[i].ticks = 0;
-            
-                if (m_Timer[i].count != 0xFFFF)
-                {
-                    m_Timer[i].count++;
-                    overflow = false;
-                }
-                else 
-                {
-                    if (m_Timer[i].interrupt) m_Interrupt->if_ |= 8 << i;
-                    m_Timer[i].count = m_Timer[i].reload;
-                    overflow = true;
-                }
-            }
-        }
-    }
-
-    ///////////////////////////////////////////////////////////
-    /// \author Frederic Meyer
-    /// \date   July 31th, 2016
-    /// \fn     RunDMA
-    ///
-    ///////////////////////////////////////////////////////////
-    void GBAMemory::RunDMA()
-    {
-        // TODO: FIFO A/B and Video Capture
-        for (int i = 0; i < 4; i++)
-        {
-            bool start = false;
-
-            if (!m_DMA[i].enable) continue;
-                
-            switch (m_DMA[i].start_time) 
-            {
-            case StartTime::Immediate:
-                start = true;
-                break;
-            case StartTime::VBlank:
-                if (m_Video->m_VBlankDMA)
-                {
-                    start = true;
-                    m_Video->m_VBlankDMA = false;
-                }
-                break;
-            case StartTime::HBlank:
-                if (m_Video->m_HBlankDMA)
-                {
-                    start = true;
-                    m_Video->m_HBlankDMA = false;
-                }
-                break;
-            case StartTime::Special:
-                #ifdef DEBUG
-                ASSERT(i == 3, LOG_ERROR, "DMA: Video Capture Mode not supported.");
-                #endif
-                break;
-            }
-            
-            m_DidTransfer = start;
-            m_DMACycles = 2;
-
-            if (start)
-            {
-                AddressControl dest_control = m_DMA[i].dest_control;
-                AddressControl source_control = m_DMA[i].source_control;
-                bool transfer_words = m_DMA[i].size == TransferSize::Word;
-
-                #if DEBUG
-                u32 value = ReadHWord(dma[i].source);
-                LOG(LOG_INFO, "DMA%d: s=%x d=%x c=%x count=%x l=%d v=%x", i, dma[i].source_int,
-                               dma[i].dest_int, 0, dma[i].count_int, video->vcount, value);
-                ASSERT(dma[i].gamepack_drq, LOG_ERROR, "Game Pak DRQ not supported.");
-                #endif
-                                        
-                // Run as long as there is data to transfer
-                while (m_DMA[i].count_int != 0)
-                {
-                    // Transfer either Word or HWord
-                    if (transfer_words)
-                    {
-                        WriteWord(m_DMA[i].dest_int & ~3, ReadWord(m_DMA[i].source_int & ~3));
-                        m_DMACycles += SequentialAccess(m_DMA[i].dest_int, AccessSize::Word) + 
-                                      SequentialAccess(m_DMA[i].source_int, AccessSize::Word);
-                    }
-                    else
-                    {
-                        WriteHWord(m_DMA[i].dest_int & ~1, ReadHWord(m_DMA[i].source_int & ~1));
-                        m_DMACycles += SequentialAccess(m_DMA[i].dest_int, AccessSize::Hword) + 
-                                      SequentialAccess(m_DMA[i].source_int, AccessSize::Hword);
-                    }
-                        
-                    // Update destination address
-                    if (dest_control == AddressControl::Increment || dest_control == AddressControl::Reload)
-                        m_DMA[i].dest_int += transfer_words ? 4 : 2;
-                    else if (dest_control == AddressControl::Decrement)
-                        m_DMA[i].dest_int -= transfer_words ? 4 : 2;
-                        
-                    // Update source address
-                    if (source_control == AddressControl::Increment || source_control == AddressControl::Reload)
-                        m_DMA[i].source_int += transfer_words ? 4 : 2;
-                    else if (source_control == AddressControl::Decrement)
-                        m_DMA[i].source_int -= transfer_words ? 4 : 2;
-                        
-                    // Update count
-                    m_DMA[i].count_int--;
-                }
-                
-                // Reschedule the DMA as specified or disable it
-                if (m_DMA[i].repeat)
-                {
-                    m_DMA[i].count_int = m_DMA[i].count & dma_count_mask[i];
-                    if (m_DMA[i].count_int == 0)
-                        m_DMA[i].count_int = dma_count_mask[i] + 1;
-                    if (dest_control == AddressControl::Reload)
-                        m_DMA[i].dest_int  = m_DMA[i].dest & dma_dest_mask[i];
-                } 
-                else 
-                {
-                    m_DMA[i].enable = false;
-                }
-                    
-                // Raise DMA interrupt if enabled
-                if (m_DMA[i].interrupt)
-                    m_Interrupt->if_ |= 256 << i;
-            }
-        }
-    }
 
     ///////////////////////////////////////////////////////////
     /// \author Frederic Meyer
